% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_S3methods.R
\name{as.character.Dictionary}
\alias{as.character.Dictionary}
\title{Create a string representation of a Dictionary}
\usage{
\method{as.character}{Dictionary}(x, n = 2, ...)
}
\arguments{
\item{x}{(\link{Dictionary})}

\item{n}{(\code{integer(1)}) \cr
Number of items to print on either side of ellipsis.}

\item{...}{(\code{ANY}) \cr Other arguments, currently unused.}
}
\description{
Creates a string representation of a \link{Dictionary} used in
printing.
}
