% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet_S3methods.R
\name{as.data.table.ParameterSet}
\alias{as.data.table.ParameterSet}
\title{Coerce a ParameterSet to a data.table}
\usage{
\method{as.data.table}{ParameterSet}(x, sort = TRUE, ...)
}
\arguments{
\item{x}{(\link{ParameterSet})}

\item{sort}{(\code{logical(1)}) \cr If \code{TRUE}(default) sorts the \link{ParameterSet}
alphabetically by id.}

\item{...}{(\code{ANY}) \cr Other arguments, currently unused.}
}
\description{
Coercion from \link{ParameterSet} to \link[data.table:data.table]{data.table::data.table}.
Dependencies, transformations, and tag properties are all lost in
coercion.
}
