% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet_S3methods.R
\name{as.ParameterSet}
\alias{as.ParameterSet}
\alias{as.ParameterSet.data.table}
\alias{as.ParameterSet.prm}
\alias{as.ParameterSet.list}
\title{Coercions to ParameterSet}
\usage{
as.ParameterSet(x, ...)

\method{as.ParameterSet}{data.table}(x, ...)

\method{as.ParameterSet}{prm}(x, ...)

\method{as.ParameterSet}{list}(x, ...)
}
\arguments{
\item{x}{(\code{ANY}) \cr Object to coerce.}

\item{...}{(\code{ANY}) \cr Other arguments passed to \link{ParameterSet}, such as
\code{tag_properties}.}
}
\description{
Coercions to ParameterSet
}
