% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm.R
\name{as.prm}
\alias{as.prm}
\alias{as.prm.ParameterSet}
\alias{as.prm.data.table}
\title{Coercion Methods to prm}
\usage{
as.prm(x, ...)

\method{as.prm}{ParameterSet}(x, ...)

\method{as.prm}{data.table}(x, ...)
}
\arguments{
\item{x}{(\code{ANY}) \cr Object to coerce.}

\item{...}{(\code{ANY}) \cr Other arguments, currently unused.}
}
\description{
Methods for coercing various objects to a \link{prm}.
}
