% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_S3methods.R
\name{c.Dictionary}
\alias{c.Dictionary}
\title{Concatenate multiple Dictionary objects}
\usage{
\method{c}{Dictionary}(...)
}
\arguments{
\item{...}{(\link{Dictionary}) \cr Dictionaries to concatenate.}
}
\description{
Creates a new \link{Dictionary} from the elements of provided
\link{Dictionary} objects, first checks keys are unique.
}
