% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet_S3methods.R
\name{c.ParameterSet}
\alias{c.ParameterSet}
\title{Concatenate ParameterSet Objects}
\usage{
\method{c}{ParameterSet}(..., pss = list(...))
}
\arguments{
\item{...}{(\link{ParameterSet}s) \cr \link{ParameterSet} objects to concatenate.}

\item{pss}{(\code{list()}) \cr Alternatively pass a list of \link{ParameterSet}
objects.}
}
\description{
Concatenate multiple \link{ParameterSet} objects into a single
\link{ParameterSet}.
}
\details{
Concatenates ids, tags, tag properties and dependencies,
but not transformations.
}
