% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{expect_equal_ps}
\alias{expect_equal_ps}
\title{Check if two parameters are equal}
\usage{
expect_equal_ps(obj, expected)
}
\arguments{
\item{obj, expected}{\link{ParameterSet}}
}
\description{
Primarily for internal use
}
