% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_S3methods.R
\name{length.Dictionary}
\alias{length.Dictionary}
\title{Get length of a Dictionary}
\usage{
\method{length}{Dictionary}(x)
}
\arguments{
\item{x}{(\link{Dictionary})}
}
\description{
This is simply a wrapper around \verb{[Dictionary]$length} to get the
number of elements in a \link{Dictionary}.
}
