% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet_S3methods.R
\name{length.ParameterSet}
\alias{length.ParameterSet}
\title{Length of a ParameterSet}
\usage{
\method{length}{ParameterSet}(x)
}
\arguments{
\item{x}{(\link{ParameterSet})}
}
\description{
Gets the number of parameters in the \link{ParameterSet}.
}
