% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{param6-package}
\alias{param6}
\alias{param6-package}
\title{param6: A Fast and Lightweight R6 Parameter Interface}
\description{
param6 makes use of {set6} alongside the S3 and R6 paradigms to provide a fast and lightweight interface for parameters and parameter sets.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk} (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})

}
