% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{pset}
\alias{pset}
\title{Convenience Function for Constructing a ParameterSet}
\usage{
pset(prms)
}
\arguments{
\item{prms}{(\code{list()}) \cr List of \link{prm} objects.}
}
\description{
See \link{ParameterSet} for full details.
}
\examples{
prms <- list(
 prm("a", Set$new(1), 1, tags = "t1"),
 prm("b", "reals", 1.5, tags = "t1"),
 prm("d", "reals", 2, tags = "t2")
)
p <- pset(prms)
}
