% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_S3methods.R
\name{[.Dictionary}
\alias{[.Dictionary}
\title{Extract an element from a Dictionary}
\usage{
\method{[}{Dictionary}(object, i)
}
\arguments{
\item{object}{(\link{Dictionary})}

\item{i}{(\code{character()}) \cr
Keys of items to get.}
}
\description{
This is simply a wrapper around \verb{[Dictionary]$get_list(i)}
}
