% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet_S3methods.R
\name{[.ParameterSet}
\alias{[.ParameterSet}
\title{Extract a sub-ParameterSet by Parameters}
\usage{
\method{[}{ParameterSet}(object, ...)
}
\arguments{
\item{object}{(\link{ParameterSet})}

\item{...}{(\code{ANY}) \cr Passed to \link{ParameterSet}$extract}
}
\description{
Creates a new \link{ParameterSet} by extracting the given
parameters. S3 method for the \verb{$extract} public method.
}
