% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary_S3methods.R
\name{summary.Dictionary}
\alias{summary.Dictionary}
\title{Summarise a Dictionary}
\usage{
\method{summary}{Dictionary}(object, n = 2, ...)
}
\arguments{
\item{object}{(\link{Dictionary})}

\item{n}{(\code{integer(1)}) \cr
Number of items to print on either side of ellipsis.}

\item{...}{(\code{ANY}) \cr Other arguments, currently unused.}
}
\description{
This is simply a wrapper around \verb{[Dictionary]$summary(n)}.
}
