% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_dictionary.R
\docType{data}
\name{support_dictionary}
\alias{support_dictionary}
\title{Support Dictionary}
\format{
An object of class \code{Dictionary} (inherits from \code{R6}) of length 31.
}
\usage{
support_dictionary
}
\description{
\link{Dictionary} for parameter supports
}
\details{
See \link{Dictionary} for full details of how to add other \link[set6:Set]{set6::Set} objects as supports
to this dictionary.
NULL
}
\keyword{datasets}
