% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet_S3methods.R
\name{[.ParameterSet}
\alias{[.ParameterSet}
\title{Extract a sub-ParameterSet by Parameters}
\usage{
\method{[}{ParameterSet}(object, i = NULL, tags = NULL, prefix = NULL, ...)
}
\arguments{
\item{object}{(\link{ParameterSet})}

\item{i}{(\code{character()}) \cr
If not \code{NULL} then specifies the parameters by id to extract. Should be
\code{NULL} if \code{prefix} is not \code{NULL}.}

\item{tags}{(\code{character()}) \cr
If not \code{NULL} then specifies the parameters by tag to extract. Should be
\code{NULL} if \code{prefix} is not \code{NULL}.}

\item{prefix}{(\code{character()}) \cr
If not \code{NULL} then extracts parameters according to their prefix and
additionally removes the prefix from the id. A prefix is determined as
the string before \code{"__"} in an id.}

\item{...}{(\code{ANY}) \cr Other arguments, currently unused.}
}
\description{
Creates a new \link{ParameterSet} by extracting the given
parameters. S3 method for the \verb{$extract} public method.
}
