% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_dictionary.R
\name{support_dictionary}
\alias{support_dictionary}
\title{Support Dictionary}
\description{
\link{Dictionary} for parameter supports
}
\details{
See \link{Dictionary} for full details of how to add other \link[set6:Set]{set6::Set}
objects as supports to this dictionary.
}
\examples{
support_dictionary$keys
support_dictionary$items
}
